Added some test code for new approaches for adding new phdrs
I believe that I got prototypes for two approaches working:
	-Shifting everything after the phdr table forward, allowing
	 the phdrs to stay at the same starting offset but expand.
	 This has the complexity that the loader still expects the
	 phdrs to be in a certain place, so the loader must be
	 modified as well (although the changes are minor).  Plus
	 the starting address of the text section is offset
	 backwards by the same amount the sections are shifted
	 forwards.  It's a bit confusing, and the prototype is
	 messy right now, with a bunch of test code that isn't
	 needed.
	-Writing the phdrs directly in the text section and
         overwriting the phdrs after the program has been
	 initialized.  This works fine if you are willing to
	 rewrite the code and insert a snippet at the start of the
	 program to patch the code back into the text section.
	 This will also require an extra section to hold the
	 overwritten bytes in the text section.
I do not have a finished prototype that actually allows these
approaches to be used in rewriting yet.  For now, they are merely
proof-of-concept, adding a few new segment/section pairs to show
that I can add multiple entries to the phdrs, but I don't have it
set up to allow for adding arbitrary new segments/sections.
I am also not sure if code for both these approaches is working
correctly in the code right now; I had to leave this code alone
for a while to work on something else, and I haven't tested it
recently.