Changed error to warning to allow for overwriting .text with phdrs
This change allows a binary with no space for additional phdrs to
move the phdrs to overwrite the start of the .text section.  This
obviously requires runtime support from the binary to restore the
original .text contents, so this library does not provide a
function to overwrite .text contents.  Rather, it simply allows
code using it as a library to relocate the phdrs anywhere, even if
the new location overlaps another segment.